Refs #200

Conflicts:
	Code/Mantid/Framework/CMakeLists.txt
	Code/Mantid/Framework/Geometry/inc/MantidGeometry/Crystal/OrientedLattice.h
	Code/Mantid/Framework/PythonInterface/plugins/algorithms/LoadSINQFile.py
	Code/Mantid/Framework/PythonInterface/plugins/algorithms/ViewBOA.py

+++ Code/Mantid/Framework/SINQ/test/LoadFlexiNexusTest.h
@@ -44,7 +44,7 @@ public:
 		IMDHistoWorkspace_sptr data =  
 		  AnalysisDataService::Instance().retrieveWS<IMDHistoWorkspace>(
 			outputSpace);
-		long nBin = data->getNPoints();
+		long nBin = static_cast<long>(data->getNPoints());
 		long sum = 0;
 		double *sdata = data->getSignalArray();
 		for(long i = 0; i < nBin; i++){
@@ -101,7 +101,7 @@ public:
 		MantidVec& X = data->dataX(0);
 		MantidVec& Y = data->dataY(0);
 		double dSum = .0;
-		for(int i = 0; i < Y.size(); i++){
+		for(size_t i = 0; i < Y.size(); i++){
 		  dSum += Y[i];
 		} 
 		TS_ASSERT_EQUALS(dSum,198812);
